Updated number of agents, starts, targets & visualization
updated the number of agents, start positions and targets. The total number of agents is input and then divided among the start positions evenly. For the targets, each agent is assigned a list of 3 random targets and then to go back to its start position, where it will be marked as done.
To get a visualization of the output the lines has been sorted into nested list, where each corresponding sub-list contains all agents movement from a step. to make sure this happened properly each agent has an id which will track it's actions and whether it has finished its list of targets.